Verify token expiry in verify endpoint

+++ src/main/java/com/pecacm/backend/services/UserService.java
@@ -14,6 +14,7 @@ import com.pecacm.backend.repository.UserRepository;
 import com.pecacm.backend.repository.VerificationTokenRepository;
 import com.pecacm.backend.response.UserEventDetails;
 import org.apache.logging.log4j.util.Strings;
+import org.springframework.cglib.core.Local;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
@@ -23,6 +24,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.*;
 
 @Service
@@ -67,7 +69,11 @@ public class UserService implements UserDetailsService {
         VerificationToken token = verificationTokenRepository.findById(tokenId).orElseThrow(() ->
                 new AcmException("Verification token not found", HttpStatus.NOT_FOUND)
         );
-        // TODO: check token expiration
+        LocalDateTime currentDate = LocalDateTime.now();
+        boolean tokenExpired = token.getCreatedDate().isBefore(currentDate.minusDays(3));
+        if (tokenExpired){
+            throw new AcmException("Token Expired", HttpStatus.NOT_FOUND)
+        }
         User user = token.getUser();
         user.setVerified(true);
         verificationTokenRepository.deleteById(tokenId);
